+++ bases/karp/karp_v6_api/routes/history_api.py
@@ -16,10 +16,10 @@ from karp.lex.application.queries import (
     GetHistory,
     GetEntryHistory,
 )
-from karp.webapp import schemas
+from karp.karp_v6_api import schemas
 
-from karp.webapp import dependencies as deps
-from karp.webapp.dependencies.fastapi_injector import inject_from_req
+from karp.karp_v6_api import dependencies as deps
+from karp.karp_v6_api.dependencies.fastapi_injector import inject_from_req
 
 
 # pylint: disable=unsubscriptable-object
