+++ karp/tests/e2e/test_resources_api.py
@@ -3,7 +3,7 @@ from starlette import status
 
 from karp import auth
 from karp.foundation.value_objects import unique_id
-from karp.webapp.schemas import ResourceCreate, ResourcePublic
+from karp.karp_v6_api.schemas import ResourceCreate, ResourcePublic
 
 
 @pytest.fixture
