+++ test/bases/karp/karp_v6_api/test_core.py
@@ -1,0 +1,5 @@
+from karp.karp_v6_api import core
+
+
+def test_sample():
+    assert core is not None
